requirements.txt
python-telegram-bot
google-generativeai
python-dotenv
fastapi
uvicorn[standard]
python-multipart